# docs/api/openapi-daily-bias.yaml
openapi: 3.0.0
info:
  title: Daily Bias Analysis API
  version: 1.0.0
  description: |
    AI-powered 6-step daily bias analysis for trading instruments
    
    ## Overview
    
    The Daily Bias Analysis API provides comprehensive market analysis through 6 sequential steps:
    1. **Security Analysis** - Volatility and risk assessment
    2. **Macro Analysis** - Economic context and events
    3. **Institutional Flux** - Volume and order flow analysis
    4. **Mag 7 Leaders** - Correlation with tech leaders
    5. **Technical Structure** - Support, resistance, and trend analysis
    6. **Synthesis** - Final bias aggregation
    
    ## Authentication
    
    All endpoints require a valid JWT token in the `Authorization` header.
    
    ## Rate Limiting
    
    - 10 requests per second (burst)
    - 1 analysis request per user per day (unless admin)
    - Cache TTL: 5 minutes per instrument
    
    ## Performance Targets
    
    - Individual steps: < 3s (p95)
    - Full analysis: < 10s (p95)
    - Cache hit: < 100ms
    
  contact:
    name: API Support
    url: https://tradingjournal.app/support
  
  license:
    name: Proprietary
    url: https://tradingjournal.app/license
servers:
  - url: https://api.tradingjournal.app
    description: Production
  - url: https://staging-api.tradingjournal.app
    description: Staging
  - url: http://localhost:3000
    description: Development

# ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token required for all endpoints
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Unauthorized"
            code: "INVALID_INPUT"
            timestamp: "2026-01-17T14:30:00Z"
...
  schemas:
...
    SentimentLevel:
      type: string
      enum: [VERY_BEARISH, BEARISH, NEUTRAL, BULLISH, VERY_BULLISH]
...
    VolumeLevel:
      type: string
      enum: [LOW, NORMAL, HIGH, EXTREMELY_HIGH]
...
    ErrorResponse:
      type: object
      required: [error, code, timestamp]
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          enum: [INVALID_INPUT, RATE_LIMIT, SERVICE_UNAVAILABLE, INTERNAL_ERROR]
        details:
          type: object
          description: Additional error context
        timestamp:
          type: string
          format: date-time
    
    # Step 1: Security Analysis
...
    EconomicEvent:
      type: object
      required: [event, time, importance, country]
      properties:
        event:
          type: string
          example: "US Non-Farm Payrolls"
        time:
          type: string
          pattern: '^\d{2}:\d{2}$'
          example: "13:30"
        importance:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        country:
          type: string
          example: "US"
        forecast:
          type: number
          nullable: true
        previous:
          type: number
          nullable: true
        actual:
          type: number
          nullable: true
        impactOnInstrument:
          type: string
    MacroAnalysis:
      type: object
      required: [economicEvents, macroScore, sentiment, timestamp, instrument]
      properties:
        economicEvents:
          type: array
          items:
            $ref: '#/components/schemas/EconomicEvent'
        macroScore:
          type: number
          minimum: 0
          maximum: 10
        sentiment:
          $ref: '#/components/schemas/SentimentLevel'
        analysis:
          type: object
          properties:
            summary:
              type: string
              maxLength: 500
            centralBankPolicy:
              type: string
            economicCycle:
              type: string
              enum: [EXPANSION, PEAK, CONTRACTION, TROUGH]
            keyThemes:
              type: array
              items:
                type: string
        timestamp:
          type: string
          format: date-time
        instrument:
          type: string
    
    # Step 3: Institutional Flux
    VolumeHeatMap:
      type: object
      required: [priceLevel, volume, intensity]
      properties:
        priceLevel:
          type: string
          example: "18500-18600"
        volume:
          type: number
          example: 125000
        intensity:
          type: number
          minimum: 0
          maximum: 1
    VolumeProfile:
      type: object
      required: [volumeLevel, trend, concentration]
      properties:
        volumeLevel:
          $ref: '#/components/schemas/VolumeLevel'
        trend:
          type: string
          enum: [DECREASING, STABLE, INCREASING]
        concentration:
          type: number
          minimum: 0
          maximum: 1
        heatMap:
          type: array
          items:
            $ref: '#/components/schemas/VolumeHeatMap'
    OrderFlow:
      type: object
      required: [buyerDominance]
      properties:
        buyerDominance:
          type: number
          minimum: 0
          maximum: 1
        largeOrders:
          type: object
          properties:
            buyOrders:
              type: integer
              minimum: 0
            sellOrders:
              type: integer
              minimum: 0
            ratio:
              type: number
              minimum: 0
        institutionalPressure:
          type: string
          enum: [BULLISH, NEUTRAL, BEARISH]
    InstitutionalFlux:
      type: object
      required: [volumeProfile, orderFlow, fluxScore, timestamp, instrument]
      properties:
        volumeProfile:
          $ref: '#/components/schemas/VolumeProfile'
        orderFlow:
          $ref: '#/components/schemas/OrderFlow'
        fluxScore:
          type: number
          minimum: 0
          maximum: 10
        analysis:
          type: object
          properties:
            summary:
              type: string
              maxLength: 500
            keyLevels:
              type: array
              items:
                type: object
                properties:
                  level:
                    type: number
                  type:
                    type: string
                    enum: [SUPPORT, RESISTANCE]
                  strength:
                    type: number
                    minimum: 0
                    maximum: 1
        timestamp:
          type: string
          format: date-time
        instrument:
          type: string
    
    # Step 4: Mag 7 Leaders
...
paths:
...
  /api/daily-bias/macro:
    post:
      summary: Macro Analysis (Step 2 only)
      operationId: getMacroAnalysis
      security:
        - BearerAuth: []
      tags:
        - Analysis Steps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instrument]
              properties:
                instrument:
                  type: string
                  example: "NQ1"
      responses:
        '200':
          description: Macro analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MacroAnalysis'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/daily-bias/flux:
    post:
      summary: Institutional Flux (Step 3 only)
      operationId: getInstitutionalFlux
      security:
        - BearerAuth: []
      tags:
        - Analysis Steps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instrument]
              properties:
                instrument:
                  type: string
                  example: "NQ1"
      responses:
        '200':
          description: Institutional flux result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstitutionalFlux'
        '401':
          $ref: '#/components/responses/Unauthorized'
...